CNN_requirements.txt
#RepresentationCNN:
tensorflow
numpy
tarfile
six.moves

#ClassificationCNN:
tensorflow
numpy
argparse
re
six.move

